[feature 141839001] Add test to check database connections - Move database tests to their own class to make code DRY
